[DataSharing] Deflake GroupDataModelTests
There is a problem with using base::Time::Now() in tests: basically the
assumption that it will always advance is wrong, it could stay the same.

It was used to populate `changed_at_...` field in specifics, which is
used by the implementation to determine whether something actually
changed.

This CL populates `changed_at_...` with an incrementing counter and this
change should deflake the test.

Bug: 372824571, 370897286
Change-Id: I9755ae6908aaa08077b8aadf2790f50ce8ab803a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5964279
Auto-Submit: Maksim Moskvitin <mmoskvitin@google.com>
Reviewed-by: Rushan Suleymanov <rushans@google.com>
Commit-Queue: Maksim Moskvitin <mmoskvitin@google.com>
Cr-Commit-Position: refs/heads/main@{#1373857}