Support nested pseudo elements styling
Currently, styling for pseudo elements is done in a way that
we have an originating element as a "base for matching" and
pseudo id to tell which pseudo element we want to collect
rules for.

But for nested pseudo elements we can't do it in such way,
since we need to have originating pseudo element as a
"base for matching".

So, this CL adds functionality for pseudo elements to be
a "base for matching", by introducing a vector of ancestor
pseudo elements to match nested pseudo elements against.

Example:
The rule is div::column::scroll-marker and we match:
  a) for div, then we won't match this rule as we can't
  reach the end of the selector;
  b) for column, the vector would be just [column],
  so we start with matching div against originating element
  (which is div), then move to ::column part and match against
  column (from vector), then we see ::scroll-marker, so we set
  ::scroll-marker pseudo style on column.
  c) for scroll marker, the vector would be [column, scroll marker],
  so we go div -> column -> scroll marker.
  d) note, that for div::before::scroll-marker we would fail matching
  at column (from vector) against ::before (from the selector), and
  the same would happen if we have scroll marker inside before inside
  div, but the rule would be div::column::scroll-marker.

Also, it changes other places that were only expecting regular element
before this CL to correctly work with a new resolve schema.

The non-existent pseudo elements are still matched with
originating element + pseudo id way, as there is no PseudoElement
to match against.

This CL also allows to create infinite ::before::before... on the
style resolver part, but it's currently blocked on the parsing site.

The test is removed from selector_checker_test as now it's fully covered
by WPT tests.

Also note that after this CL authors can now style nested markers.

Change-Id: If21a2a4158ebb7a63cebab8a33e5c63f0cd38323
Bug: 373478544
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5891553
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373874}