[iOS][Account Menu]Fix double tap issue
This fixes a crash in M131, and I hope to cherry-pick it.

In the NTP, a double tap on the identity disc could cause a crash by requesting to open the account menu twice.

Similarly, on ipad, when the popover is closing, but the coordinator not yet fully stopped, tapping on the account menu could cause a crash.

Those are now solved.

Technically, this crash could also occur if the user tapped on other
buttons in the same scene that would open a signin coordinator. Luckily, there are no such buttons today on the NTP, and opening a new NTP would be slow enough to avoid the race condition issue entirely.

Fixed: 365052502
Change-Id: I668b1223888abc8d90b825d1cde641c50d05b95c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5961324
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373879}