FSA: Use the profile path for base::DIR_HOME on ChromeOS
Updates the FSA permission context's block list logic to use the profile
path for base::DIR_HOME on ChromeOS.

ChromeOS supports multi-user sign-in. base::DIR_HOME only gives the home
directory of the primary account, the first user to sign in. This
updates the block list logic so that it always uses the profile path
of the BrowserContext which ChromeOS uses as the home directory.

Fixed: 367429476
Change-Id: I1fa652195ca42b60be685755e6182df7cb2a42c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5963016
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Daseul Lee <dslee@chromium.org>
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373955}