Introduce SearchActivity IntentBuilder.
This change replaces a growing number of mostly-optional parameters
and functions performing similar tasks with an IntentBuilder, allowing
the caller to directly specify the kind of interaction they need with
the SearchActivity without passing a lithany of parameters.

The result permits the caller to rely on SearchActivity's built-in
mechanisms such as DSE URL computation, or dispatching Open intents to
Chrome.

Bug: b:370570081
Change-Id: I67528da8a4e256acc57b253b782e5bb2bd3b62d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5963321
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373997}