[Pwa/nav Capturing] Fix parameters for flaky tests
When the CaptureOn and CaptureOff tests were split, the list of disabled
flaky tests were not updated, leading to test params that were no
longer disabled.

This CL fixes the params so that the tests are still disabled. This is
fine for the time being since these params aren't duplicated across
CaptureOn and CaptureOff tests.

Bug: 372119276
Change-Id: If42d1795a56795593fefd8d1fa29709b848661de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5963805
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374027}