Set tree priority to NEW_CONTENT when scroll checkerboards
This logic was removed in CL:5565398 due to obsolete logic which was
used to determine whether a page checkerboarded due to no recording.
Since then, paint checkerboarding tracking has been updated to work
post composite-after-paint, and therefore this change can be
reintroduced.

The optimization in this CL allows CC to attempt to block on the main
thread in the event a scroll engenders checkerboarding. This is done by
setting the tree priority to NEW_CONTENT_TAKES_PRIORITY, which
deprioritizes impl latency in the CC scheduler.

Bug: 40680264
Change-Id: I19659441bd2a8dc86685af4f6b1afcfaeec4e47d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5916849
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1374055}