Rename AuctionProcessManagerTests.
In particular:

* AuctionProcessManagerTests -> SitePerProcessAuctionProcessManagerTest
* InRendererAuctionProcessManagerTest_NoOriginKeyedProcessesByDefault ->
    SharedRendererInRendererAuctionProcessManagerTest

Also remove the test fixture of the latter, which was enabling a
feature which the base AuctionProcessManagerTest already enables,
albeit with a different timeout that those tests don't exercise,
anyways.

This CL does not rename the AuctionProcessManagerTest fixture itself,
which is now aliased to the aforementioned two renamed sets of tests.

Bug: 374253381
Change-Id: If539132e94efa08bbb3f8a35c7ad041d9008bf63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5966414
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374065}