exception-db: Add one more retry for the exception record upload.
The code used to only try two times, and report the upload failure in
the last try. This CL increases the try times to 3 so that it will try
to clear the stacktrace and upload in the 2nd try. This will help
preserve more exception data records.

Also add unittest for the exception data upload.

Bug: 341362003
Change-Id: I3d924be3ee072bd5d49c44527e1a1d6648387e79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5953815
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1374101}