[FaceGaze] Exclude blink when detecting eye squint left/right
* Add logic to GestureDetector to ignore certain instances of
gestures if certain MediapipeFacialGestures are detected at
the same time at a confidence of > 0.5.
* For EYE_SQUINT_LEFT and EYE_SQUINT_RIGHT, if an eye squint
or blink on the other side is also detected and at > 0.5,
do not register the gesture as the user is likely blinking
involuntarily.
* Add test to check that blinking will not trigger
EYE_SQUINT_LEFT or EYE_SQUINT_RIGHT.

AX-Relnotes: N/A
Fixed: 374774499
Change-Id: I937588c4ac1246e428b86b430a77fa58aa4c6d79
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5962467
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Amanda Lin Dietz <aldietz@google.com>
Cr-Commit-Position: refs/heads/main@{#1374180}