[E2E] Fix bottom sheet over translate when edge to edge
Under E2E, bottom sheets contents are pushed out from the nav bar region
by having an added translationY (introduced in crrev.com/c/5239157).
When the bottom sheet is in the full height, the translation can turn
into negative value. One example is the AllPasswordBottomSheetView.

This CL limits the translationY to be not less than 0. In the meantime,
ensure the bottom padding for the sheet content is the size of the
bottom insets so no bottom sheet content is drawn under the nav bar.

https://screencast.googleplex.com/cast/NjY3Mjg1MzkwNTcwMjkxMnw1NTRkMzVkMy01Yg

Bug: 375212748
Change-Id: I75605178a95dd01f67f43c4f4d932e9bf4c343ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5962738
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374205}