[WebAuthn] Unregister enclave client when using unknown key
We have seen issues in production with users having client keys that
are supposedly registered but are unknown to the enclave
authenticator service.

It is possible that lower-level communication failures are resulting
in that state on Windows during deferred UV key creation. This change
causes `Transact` to provide more detailed errors to its callback,
rather than a null pointer.

Also in this change, EnclaveAuthenticator will notice when the
service returns a missing key response, and unregister the client so
the user can try to re-register on their next GPM passkey use attempt.

Metrics on these errors will be coming in another change soon so we
can further investigate causes.

Bug: 375211367
Change-Id: I26335752b949f67230f06292702efd1a3d2e0cf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5962740
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374220}