[webui] Introduce utilities to manage BrowserWindowInterface
WebUI clients will be able to retrieve and listen for changes to
their embedding BrowserWindowInterface.

It is the responsibility of embedders (typically the WebUI
embedding frameworks such as the WebUI Bubble and Side Panel
infrastructure) to call SetBrowserWindowInterface() when the
browser associated with a WebContents they manage changes.

This CL only adds the utilities and relevant tests. See
downstream CLs for an example of how this will be used practically.

See go/webui-embedding-context for details.

Bug: 360724768
Change-Id: I8f5f8859907084093f93b1d965c907b612e87a37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5960779
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374418}