lobster: fix insertion logic
Insertion should only happen when the original input field is refocused.
Currently, insertion does not work because it is triggered when the
webui is still present.

To make insertion work, we follow the approach that editor uses:
* Insertion triggered by webui
* Inflate the *data*
* Cache the inflated data, using a new class called `LobsterInsertion`.
* Close the Webui. This refocuses into the original input field.
* Attempt the insertion.

Noted that the webui needs to receive the status before it is closed,
otherwise it will cause crashes. In this case, the status of the request
to inflate the image is used to return back to webui.

Bug: b:374682805
Change-Id: Ie3b48037bd38261fca9df0293559602686015fd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5960923
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Commit-Queue: Chuong Ho <hdchuong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374468}