lobster: add insert/download mode
Currently, Lobster UI always shows in Download mode. In reality, if the
input field is a richly editable field, then it should prioritize insert
mode over download mode.

This CL adds a logic for Lobster to be able to determine when to
show in Download/Insert mode. The essence lives in the bool value that
TextInputClient::CanInsertImage returns. This boolean can be passed as
a param into webui to show the correct mode.

For picker entry point, CanInsertImage signal of the current focused
text input field must be retrieved before the Picker UI shows up. This
CL also does some refactoring for the following:

* Create the Lobster Trigger at the time the Quick Insert widget shows
up.

* Cache the Lobster Mode into the Lobster Trigger and pass the mode to
Lobster Session when the session is created.

Bug: b:373733364
Change-Id: I2be9ccc3babf5dbafc2e1314a0b43ddf69034e8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5963879
Commit-Queue: Chuong Ho <hdchuong@chromium.org>
Reviewed-by: Curtis McMullan <curtismcmullan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374473}