Match prefetches against render requests
Call the matching code from PrefetchMatchingURLLoaderFactory. If the
newly-added "NetworkContextPrefetchUseMatches" feature is enabled, also
allow successful matches to be consumed by the renderer. Otherwise, only
the histograms indicating matching success or failure will be logged.
The prefetch is abandoned after a delay but the renderer may still
be able to use the disk cache entry it created.

Add check that the `request_initiator` is valid so that the render
process cannot bypass access checks when a prefetch is matched.

Rename the SetClient() method of network::PrefetchURLLoaderClient to
Consume() and make it take a PendingReceiver<URLLoader> argument. The
CorsURLLoader that the client is bound to is rebound to the passed-in
argument, permitted it to be called from the render process directly.
Add a Matches() method which calls PrefetchMatches() against the stored
ResourceRequest.

Also refactor the PrefetchURLLoaderClient tests to provide a more
realistic environment.

Currently tests that call Consume() leak memory and so are disabled on
sanitizer builds. See https://crbug.com/375425822. This memory leak only
happens if the "NetworkContextPrefetchUseMatches" feature is enabled and
will be fixed before that is used in production.

BUG=342445996

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL Uncovered code has no callers yet.
Change-Id: I607db0dbbf8017730b56de3baa4d66d2b15268ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5715614
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374542}