[Signin] Ignore ManageAccounts header on background tabs
Prior to this CL, ProcessMirrorHeader was accepting
X-Chrome-Manage-Accounts header from all WebContents, even if that
WebContents was in a background tab. Because of that, background
tabs could trigger account management screen and other parts of the
Mirror protocol. This CL adds a flag-guarded check that ensures that
the corresponding WebContents is actually in the foreground.

Bug: 372093832
Change-Id: Ib8675254e55087f113605cbf322bd546437684d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5957787
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374570}