[windows] Include the return value of CoInitializeEx in crashes
WorkerThreadCOMDelegate fails to inititialize the thread as a COM
single-threaded apartment from time to time. Many of these are due to
OOM or ATOM exhaustion, and are handled by Chrome's
`OnResourceExhausted` handler, but some are not. Include the return
value from CoInitializeEx in a crash key in this case so that we can try
to gain a bit more understanding of what is actually happening here.

Bug: 40074523
Change-Id: Ie9b305ed7227cd9c54f76d4e79f9f38ca4a88dc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5970334
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374581}