[ios][af] Throttle form extraction requests for scanning documents
Throttle the unfiltered form extraction requests used for scanning
forms in a document. Scanning should usually not cause hanging
where it is meant to be used for taking a snapshot of the forms in
the document from time to time.

For instance, this allows throttling the extraction requests
triggered by the xframe form forest.

Does not throttle the filtered form extraction requests (aka on
the spot) that are used to get form data for a particular form
which can sometimes block transactions flows (e.g. the
suggestions pipeline) and can have a user visible impact.

Hence, the feature here only targets throttling scans.

Another feature will be used to throttle on the spot form fetching.
in follow-up CL.

Change-Id: Ie982fce74c22832d8bec3519cadeb11636064e18
Bug: 372723354
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5938914
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1374592}