[DNT] Continuing animation when a second navigation commits mid gesture
With this change when navigating from B to A, if a new navigation
commits (C), the live page B will be replaced by C and the animation
continues navigating from C to A as expected. This way we are navigating
to A which is consistent with the screenshot that the user is seeing and
we're also handling the navigation to C by replacing the live page with
it.

Bug: 366023026
Change-Id: I3f60396b84630c16e6a467aa6ba6ec917053536c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5935055
Commit-Queue: Baran Erfani <baranerf@google.com>
Reviewed-by: William Liu <liuwilliam@chromium.org>
Reviewed-by: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/main@{#1374628}