[N&C] Remove LINT bounds
This CL removes the LINT bounds in the privacy sandbox service and
notice service. We re-evaluated conditions under which both services
are needed. Privacy sandbox service needs to be created for a variety
of profiles, but the notice service only needs to be created for
regular profiles.

Bug: b:372479795

Change-Id: Ia094447edf78e328590ab4f0ffd2aad6975ee56e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5921317
Commit-Queue: Aashna Sheth <aashnas@google.com>
Reviewed-by: Christine Ying <chrstne@google.com>
Reviewed-by: Abe Boujane <boujane@google.com>
Cr-Commit-Position: refs/heads/main@{#1374642}