Use a single method to request proxy config refresh
Previously, `RefreshProxyListForGeoChange` was specifically for geo
changes, and did nothing if tokens are not cached per-geo. If the
minimum time had not elapsed, it scheduled a refresh for after that
minimum time, and otherwise performed the refresh immediately.

`RequestRefreshProxyList` refreshed if the minimum time had elapsed,
otherwise it did nothing.

The only call to `RefreshProxyListForGeoChange` already checks whether
tokens are cached per-geo, so that conditional was unnecessary. And, the
behavior of scheduling a refresh after the minimum time has elapsed is
always useful.

So, this CL combines the two methods into one.

Bug: 375253808
Change-Id: Ib125254e0b1befc81243fd154736b78927b23768
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5958196
Reviewed-by: Abhi Patel <abhipatel@chromium.org>
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374664}