Only ensure _successful_ proxy config fetches are not too frequent
This check is to ensure that when signals suggest the proxy config is
out of date (such as mismatched geos), we don't constantly try to update
it. Handling for backoff on error is done in the fetcher. So when
there's an error, we can fallback to that backoff.

This also fixes the case where, on startup, the user is not signed in
(causing proxy config fetch to fail), but sign-in occurs before the
minimum time. This is only an issue when using OAuth tokens for
ProxyConfig, which is only done for debugging/testing purposes.

Bug: 375253808
Change-Id: I1bc05037293e2e3c4d25e0d749e64a1c19662127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5967337
Commit-Queue: Dustin Mitchell <djmitche@chromium.org>
Reviewed-by: Abhi Patel <abhipatel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374689}