Use ThenCallable<> for promise handling in navigator_auction.cc
ThenCallable<> handles type checking and conversion, so a lot less
hand-written logic is needed to process the resolve value of a
promise.

ThenCallable<> is templated with, among other things, the type of
the promise being observed. This does not play nicely with the
existing class hierarchy in naivgator_auction.cc. Make
AuctionHandlerFunction a GarbageCollectedMixin instead of being in
the direct class hierarchy, and use a virtual Attach() function on
it to handle actually starting listening to the promise result.

Change-Id: I0cabbd43c80fc21b17a5c1c7f09a8586a2e1bd9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5954744
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374738}