Update TextCluster and fillTextCluster() IDL definitions
Following the discussion on the WhatWG issue for Enhanced TextMetrics,
it was agreed to make all attributes of the TextCluster interface
readonly, and remove the text attribute from the JS API, which will
still be stored by the underlying C++ object. This CL brings these
changes to the TextCluster IDL interface.

Additionally, it was also decided to make the X and Y parameters for
ctx.fillTextCluster() mandatory. This CL updates its IDL definition to
reflect this.

Bug: 341213359
Change-Id: I10e2a6a7f9ece68b53f07df595006f46f464452b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939086
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374757}