media/gpu/chromeos/frame_resource: Add tracking token
This is intended to be used as a substitute for the
GenericSharedMemoryId that is returned by GetSharedMemoryId(). For new
frames, the UnguessableToken is generated. When creating a wrapping
frame, the original UnguessableToken is reused.

The tracking token is stored in the VideoFrameMetadata.

For NativePixmapFrameResource, it is created at construction time. When
a frame is wrapped, the metadata is copied to the new wrapping frame.
When a user calls set_metadata(), the original tracking token is
persisted.

For VideoFrameResource, the tracking token is passed through from the
underlying VideoFrame's VideoFrameMetadata.

Bug: b:372695223
Change-Id: I1eb674ed33cc11b6572f13c58a6858b02077c5cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5924234
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1374758}